Refactor: Move wait_until from _stbt.core to _stbt.wait
I'm moving non-GStreamer stuff out of _stbt.core.
